Better method of sizing
In practice, the task to get the size of the input stream is generally
different than the task to get the input stream itself, and may be
accessed through different systems.

Given that getting the length is supposed to be cheap and fast, you
really want to compute it before giving the user their input.

Given that this interface works well as a delegate, you then don't
want to unnecessarily precompute the length at a lower level if it
can be more cheaply computed at a higher level.

This PR adds SizedSeekableDataInput, which adds the length method
to SeekableDataInput.

The lower level implementer can then implement SeekableDataInput,
and at any level we can wrap with SizingSeekableDataInput in order to
provide a length, computing by any means necessary.